model_app/requirements.txt
Flask==1.1.1
gunicorn==19.9.0
google-cloud-storage==1.2.0
nltk==3.5
gensim==3.8.3
flair==0.5
razdel==0.4.0
